Fixes for the patcher
Testing, fixes, and validation for the patcher
